backend/requirements.txt
django
Pillow
gunicorn
#configparser
#psycopg2-binary
asgiref
djangorestframework
pytz
sqlparse
markdown
django-filter
drf-yasg[validation]
drf-spectacular
coreapi
djangorestframework-jsonapi[django-filter]
djangorestframework-jsonapi[django-polymorphic]
djangorestframework-jsonapi[openapi]
django-oauth-toolkit
django-cors-middleware
djangorestframework-simplejwt
django-rest-passwordreset
django-sendgrid-v5
djoser
social-auth-app-django
django-override-settings
django-templated-mail